refactor: consolidate DOM transform utilities
