--- svg-time-series/src/utils/domNodeTransform.ts
@@ -1,5 +1,0 @@
-export function updateNode(n: SVGGraphicsElement, m: DOMMatrix | SVGMatrix) {
-  const svgTransformList = n.transform.baseVal;
-  const t = svgTransformList.createSVGTransformFromMatrix(m);
-  svgTransformList.initialize(t);
-}
