+++ svg-time-series/src/viewZoomTransform.ts
@@ -2,5 +2,5 @@ export {
   applyAR1ToMatrixX,
   applyAR1ToMatrixY,
   applyDirectProductToMatrix,
-} from "./utils/domMatrix.ts";
-export { updateNode } from "./utils/domNodeTransform.ts";
+  updateNode,
+} from "./utils/transform.ts";
